example: add caching example
We don't really *need* (we may want to, but no need)
to recreate and reload everything at every request.
Actually, the loading and the compilation stage can be
done ahead of time, and only the instantiation can be
done per-request. We can probably pool the functions
(xref: https://github.com/http-wasm/http-wasm-host-go)
but re-instantiating (and only that) per-request
at least makes sense.

To get there, we decompose the handler into loader
and engine. Besides this decomposition, the only addition
is full io (in, out, err) wiring to the wasm module.

Signed-off-by: Francesco Romani <fromani@gmail.com>