example: add host functions example
add example with hostfunctions to enable instance reuse.
Reusing instance meaning we need (or at least there's a strong
case for) i/o functions; the other alternative is a smarter
handling of stdin/stdout/stderr.

While in a real world case the stdio route should not be easily
discarded, for the learning purpose we take the chance to
add and use host functions.

Signed-off-by: Francesco Romani <fromani@gmail.com>